Move pipeline out of the helper package 🌿
