+++ pkg/cmd/pipeline/logs.go
@@ -23,8 +23,8 @@ import (
 	"github.com/tektoncd/cli/pkg/cmd/pipelinerun"
 	"github.com/tektoncd/cli/pkg/flags"
 	"github.com/tektoncd/cli/pkg/options"
-	"github.com/tektoncd/cli/pkg/helper/pipeline"
-	phelper "github.com/tektoncd/cli/pkg/helper/pipeline"
+	"github.com/tektoncd/cli/pkg/pipeline"
+	phelper "github.com/tektoncd/cli/pkg/pipeline"
 	prhelper "github.com/tektoncd/cli/pkg/pipelinerun"
 	validate "github.com/tektoncd/cli/pkg/validate"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
