+++ pkg/pipeline/pipeline_test.go
